requirements.txt
google-generativeai>=0.3.0
opencv-python>=4.7.0
PyQt6>=6.4.0
duckduckgo-search>=3.0.0
requests>=2.28.1
beautifulsoup4>=4.11.1
edge-tts>=6.1.0
python-dotenv>=0.21.0
pillow>=9.3.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
numpy>=1.22.0 